Move span stuff to own module
